+++ lazy_sign_in/lib/lazy_sign_in.dart
@@ -2,4 +2,4 @@ library lazy;
 
 export 'src/sign_in.dart' show SignIn;
 export 'src/sign_in_ext_dummy.dart' show SignInDummy;
-export 'src/sign_in_msg.dart' show SignInMsg;
+export 'src/is_signed_in.dart' show IsSignedIn;