+++ lazy_sign_in/lib/src/sign_in_msg.dart
@@ -1,8 +1,0 @@
-/// ### Lazy [SignInMsg]
-/// - [status] == ([token].isNotEmpty)
-/// - [token] : access token
-class SignInMsg {
-  bool get status => (token.isNotEmpty);
-  final String token;
-  SignInMsg({this.token = ''});
-}